+++ tests/test_main.py
@@ -45,11 +45,7 @@ def test_main_noargs(mocker):
 
 def test_main_sys_argv(tomcat_manager_server, capsys, mocker, monkeypatch):
     # don't let it load a config file
-    mocker.patch(
-        "tomcatmanager.InteractiveTomcatManager.config_file",
-        new_callable=mock.PropertyMock,
-        return_value=None,
-    )
+    mocker.patch("tomcatmanager.InteractiveTomcatManager.load_config", autospec=True)
 
     # hack up sys.argv
     monkeypatch.setattr(
@@ -77,10 +73,8 @@ def test_main_sys_argv(tomcat_manager_server, capsys, mocker, monkeypatch):
 
 def test_main_user_password_url_command(tomcat_manager_server, mocker, capsys):
     # don't let it load a config file
-    mocker.patch(
-        "tomcatmanager.InteractiveTomcatManager.config_file",
-        new_callable=mock.PropertyMock,
-        return_value=None,
+    loader = mocker.patch(
+        "tomcatmanager.InteractiveTomcatManager.load_config", autospec=True
     )
 
     cmdline = (
@@ -92,6 +86,9 @@ def test_main_user_password_url_command(tomcat_manager_server, mocker, capsys):
     out, err = capsys.readouterr()
     out = out.splitlines()
     err = err.splitlines()
+
+    # make sure it tried to load the config file
+    assert loader.called
     assert exit_code == 0
     assert "Path" in out[0]
     assert "Sessions" in out[0]
@@ -100,11 +97,7 @@ def test_main_user_password_url_command(tomcat_manager_server, mocker, capsys):
 
 def test_main_quiet(tomcat_manager_server, mocker, capsys):
     # don't let it load a config file
-    mocker.patch(
-        "tomcatmanager.InteractiveTomcatManager.config_file",
-        new_callable=mock.PropertyMock,
-        return_value=None,
-    )
+    mocker.patch("tomcatmanager.InteractiveTomcatManager.load_config", autospec=True)
 
     cmdline = (
         f"-q -u {tomcat_manager_server.user}"
@@ -139,13 +132,42 @@ def test_main_version(capsys):
     assert not err
 
 
+def test_main_noconfig(tomcat_manager_server, capsys, mocker, monkeypatch):
+    # mock the config loader so we can check if it was called
+    loader = mocker.patch(
+        "tomcatmanager.InteractiveTomcatManager.load_config", autospec=True
+    )
+
+    # hack up sys.argv
+    monkeypatch.setattr(
+        "sys.argv",
+        [
+            "tomcat-manager",
+            "--noconfig",
+            "-u",
+            tomcat_manager_server.user,
+            "-p",
+            tomcat_manager_server.password,
+            tomcat_manager_server.url,
+            "list",
+        ],
+    )
+
+    exit_code = main()
+    out, err = capsys.readouterr()
+    out = out.splitlines()
+    err = err.splitlines()
+
+    assert not loader.called
+    assert exit_code == 0
+    assert "Path" in out[0]
+    assert "Sessions" in out[0]
+    assert "connected to" in err[0]
+
+
 def test_main_debug(tomcat_manager_server, mocker, capsys):
     # don't let it load a config file
-    mocker.patch(
-        "tomcatmanager.InteractiveTomcatManager.config_file",
-        new_callable=mock.PropertyMock,
-        return_value=None,
-    )
+    mocker.patch("tomcatmanager.InteractiveTomcatManager.load_config", autospec=True)
 
     cmdline = (
         f"-d -u {tomcat_manager_server.user}"
@@ -163,11 +185,7 @@ def test_main_debug(tomcat_manager_server, mocker, capsys):
 
 def test_main_version_with_others(tomcat_manager_server, mocker, capsys):
     # don't let it load a config file
-    mocker.patch(
-        "tomcatmanager.InteractiveTomcatManager.config_file",
-        new_callable=mock.PropertyMock,
-        return_value=None,
-    )
+    mocker.patch("tomcatmanager.InteractiveTomcatManager.load_config", autospec=True)
 
     cmdline = (
         f"-v -q -u {tomcat_manager_server.user}"
@@ -185,11 +203,7 @@ def test_main_version_with_others(tomcat_manager_server, mocker, capsys):
 
 def test_main_stdin(tomcat_manager_server, mocker, capsys):
     # don't let it load a config file
-    mocker.patch(
-        "tomcatmanager.InteractiveTomcatManager.config_file",
-        new_callable=mock.PropertyMock,
-        return_value=None,
-    )
+    mocker.patch("tomcatmanager.InteractiveTomcatManager.load_config", autospec=True)
 
     inio = io.StringIO("list\n")
     stdin = sys.stdin
@@ -215,11 +229,7 @@ def test_main_stdin(tomcat_manager_server, mocker, capsys):
 
 def test_main_echo(tomcat_manager_server, mocker, capsys):
     # don't let it load a config file
-    mocker.patch(
-        "tomcatmanager.InteractiveTomcatManager.config_file",
-        new_callable=mock.PropertyMock,
-        return_value=None,
-    )
+    mocker.patch("tomcatmanager.InteractiveTomcatManager.load_config", autospec=True)
 
     inio = io.StringIO("list\n")
     stdin = sys.stdin
@@ -246,11 +256,7 @@ def test_main_echo(tomcat_manager_server, mocker, capsys):
 
 def test_main_status_to_stdout(tomcat_manager_server, mocker, capsys):
     # don't let it load a config file
-    mocker.patch(
-        "tomcatmanager.InteractiveTomcatManager.config_file",
-        new_callable=mock.PropertyMock,
-        return_value=None,
-    )
+    mocker.patch("tomcatmanager.InteractiveTomcatManager.load_config", autospec=True)
 
     cmdline = (
         f"-s -u {tomcat_manager_server.user}"
@@ -270,11 +276,7 @@ def test_main_status_to_stdout(tomcat_manager_server, mocker, capsys):
 
 def test_main_timeout(tomcat_manager_server, mocker, capsys):
     # don't let it load a config file
-    mocker.patch(
-        "tomcatmanager.InteractiveTomcatManager.config_file",
-        new_callable=mock.PropertyMock,
-        return_value=None,
-    )
+    mocker.patch("tomcatmanager.InteractiveTomcatManager.load_config", autospec=True)
 
     cmdline = (
         f"-t 7.8 -u {tomcat_manager_server.user}"
@@ -291,11 +293,7 @@ def test_main_timeout(tomcat_manager_server, mocker, capsys):
 
 def test_main_timeout_zero(tomcat_manager_server, mocker, capsys):
     # don't let it load a config file
-    mocker.patch(
-        "tomcatmanager.InteractiveTomcatManager.config_file",
-        new_callable=mock.PropertyMock,
-        return_value=None,
-    )
+    mocker.patch("tomcatmanager.InteractiveTomcatManager.load_config", autospec=True)
 
     cmdline = (
         f"-t 0 -u {tomcat_manager_server.user}"